+++ chaosazure/webapp/probes.py
@@ -1,7 +1,7 @@
 from chaoslib import Configuration, Secrets
 from logzero import logger
 
-from chaosazure.rgraph.resource_graph import fetch_resources
+from chaosazure.common.resources.graph import fetch_resources
 from chaosazure.webapp.constants import RES_TYPE_WEBAPP